Fixing casting issues

+++ app/src/main/java/com/modarb/android/ui/home/ui/nutrition/domain/models/daily_goals/Data.kt
@@ -1,8 +1,8 @@
 package com.modarb.android.ui.home.ui.nutrition.domain.models.daily_goals
 
 data class Data(
-    val exercisesCals: Int,
-    val exercisesHours: Int,
+    val exercisesCals: Double,
+    val exercisesHours: Double,
     val sleepDone: Int,
     val sleepGoal: Int,
     val stepsDone: Int,
